+++ src/Makefile.am
@@ -261,6 +261,7 @@ BITCOIN_CORE_H = \
   torcontrol.h \
   txdb.h \
   txmempool.h \
+  txmempool_entry.h \
   txorphanage.h \
   txrequest.h \
   undo.h \
@@ -425,6 +426,7 @@ libbitcoin_node_a_SOURCES = \
   torcontrol.cpp \
   txdb.cpp \
   txmempool.cpp \
+  txmempool_entry.cpp \
   txorphanage.cpp \
   txrequest.cpp \
   validation.cpp \
@@ -930,6 +932,7 @@ libbitcoinkernel_la_SOURCES = \
   threadinterrupt.cpp \
   txdb.cpp \
   txmempool.cpp \
+  txmempool_entry.cpp \
   uint256.cpp \
   util/check.cpp \
   util/getuniquepath.cpp \
