+++ src/node/interfaces.cpp
@@ -37,6 +37,7 @@
 #include <support/allocators/secure.h>
 #include <sync.h>
 #include <txmempool.h>
+#include <txmempool_entry.h>
 #include <uint256.h>
 #include <univalue.h>
 #include <util/check.h>
