+++ src/policy/rbf.cpp
@@ -10,6 +10,7 @@
 #include <sync.h>
 #include <tinyformat.h>
 #include <txmempool.h>
+#include <txmempool_entry.h>
 #include <uint256.h>
 #include <util/moneystr.h>
 #include <util/rbf.h>
