+++ src/test/util/setup_common.cpp
@@ -39,6 +39,7 @@
 #include <timedata.h>
 #include <txdb.h>
 #include <txmempool.h>
+#include <txmempool_entry.h>
 #include <util/strencodings.h>
 #include <util/string.h>
 #include <util/thread.h>
